refactor: 기존 dto inner class 로 정리
