+++ src/main/java/org/deco/gachicoding/post/question/presentation/QuestionController.java
@@ -3,12 +3,8 @@ package org.deco.gachicoding.post.question.presentation;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.deco.gachicoding.post.question.QuestionDto;
+import org.deco.gachicoding.post.question.application.dto.QuestionDto;
 import org.deco.gachicoding.post.question.application.QuestionService;
-import org.deco.gachicoding.post.question.presentation.dto.QuestionAssembler;
-import org.deco.gachicoding.post.question.presentation.dto.response.QuestionListResponse;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -16,7 +12,6 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.net.URI;
-import java.util.List;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -39,20 +34,15 @@ public class QuestionController {
     }
 
 
-    // 이거 키워드만 받으면 안되남?
-    @ApiOperation(value = "질문 리스트")
-    @GetMapping("/question/list")
-    public ResponseEntity<List<QuestionListResponse>> getQuestionList(
-            @RequestParam(value = "keyword", defaultValue = "") String keyword,
-            @PageableDefault(size = 10) Pageable pageable
-    ) {
-
-        return ResponseEntity.ok(
-                QuestionAssembler.questionListResponse(
-                        questionService.getQuestionList(keyword, pageable)
-                )
-        );
-    }
+    // 이거 키워드만 받으면 안되나?
+//    @ApiOperation(value = "질문 리스트")
+//    @GetMapping("/question/list")
+//    public List<QuestionDto.ListResponseDto> getQuestionList(
+//            @RequestParam String keyword,
+//            @PageableDefault(size = 10) Pageable pageable
+//    ) {
+//        return QuestionAssembler.questionListResponse(questionService.getQuestionList(keyword, pageable));
+//    }
 
     @ApiOperation(value = "질문 디테일")
     @GetMapping("/question/{queIdx}")
@@ -63,7 +53,7 @@ public class QuestionController {
     }
 
     @ApiOperation(value = "질문 수정")
-    @PreAuthorize("isAuthenticated() and #request.userEmail eq principal.username")
+    @PreAuthorize("isAuthenticated() and (#request.userEmail eq principal.username)")
     @PatchMapping("/question/modify")
     public ResponseEntity modifyQuestion(
             @RequestBody @Valid QuestionDto.UpdateRequestDto request
@@ -77,7 +67,7 @@ public class QuestionController {
     }
 
     @ApiOperation(value = "질문 비활성화")
-    @PreAuthorize("isAuthenticated() and userEmail eq principal.username")
+    @PreAuthorize("isAuthenticated() and (userEmail eq principal.username)")
     @PatchMapping("/question/disable")
     public Long disableQuestion(
             @RequestParam Long queIdx,
@@ -87,7 +77,7 @@ public class QuestionController {
     }
 
     @ApiOperation(value = "질문 활성화")
-    @PreAuthorize("isAuthenticated() and userEmail eq principal.username")
+    @PreAuthorize("isAuthenticated() and (userEmail eq principal.username)")
     @PatchMapping("/question/enable")
     public Long enableQuestion(
             @RequestParam Long queIdx,
@@ -97,7 +87,7 @@ public class QuestionController {
     }
 
     @ApiOperation(value = "질문 삭제")
-    @PreAuthorize("isAuthenticated() and userEmail eq principal.username")
+    @PreAuthorize("isAuthenticated() and (userEmail eq principal.username)")
     @DeleteMapping("/question")
     public void removeQuestion(
             @RequestParam Long queIdx,
